+++ src/clients/preferences/create/create.spec.ts
@@ -1,7 +1,8 @@
 import create from '.';
 import { RestClient } from '@utils/restClient';
 import { MercadoPagoConfig } from '@src/mercadoPagoConfig';
-import type { PreferenceRequest } from './types';
+import type { PreferenceRequest } from '@src/clients/preferences/commonTypes';
+
 
 jest.mock('@utils/restClient');
